+++ internal/auth/vk_http_mock.go
