+++ internal/db/db.go
@@ -1,16 +1,17 @@
 package db
 
 import (
-	"fmt"
+	"net"
+	"strconv"
 
-	"github.com/FreakyGranny/launchpad-api/internal/app/config"
+	"github.com/FreakyGranny/launchpad-api/internal/config"
 	"github.com/go-pg/pg/v10"
 )
 
 // Connect ...
 func Connect(cfg *config.PgConnection) (*pg.DB, error) {
 	db := pg.Connect(&pg.Options{
-		Addr:     fmt.Sprintf("%s:%v", cfg.Host, cfg.Port),
+		Addr:     net.JoinHostPort(cfg.Host, strconv.Itoa(cfg.Port)),
 		User:     cfg.Username,
 		Password: cfg.Password,
 		Database: cfg.DbName,
