+++ internal/handlers/category_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/suite"
 
-	"github.com/FreakyGranny/launchpad-api/internal/app/mocks"
-	"github.com/FreakyGranny/launchpad-api/internal/app/models"
+	"github.com/FreakyGranny/launchpad-api/internal/mocks"
+	"github.com/FreakyGranny/launchpad-api/internal/models"
 )
 
 type CategorySuite struct {
